UiImpl.java
Moved UiImpl to a new file as discussed. Changes to interface include new methods to simplify switch statement. Call Methods and Constructor
