+++ bank-app/src/main/java/com/bank/exceptions/MenuException.java
@@ -3,6 +3,7 @@ package com.bank.exceptions;
 public class MenuException extends Exception {
 
     public MenuException(String string) {
+        
     }
 
 }
